+++ src/main/java/com/kiwi/observability/metrics/ThreadPoolMetrics.java
@@ -9,5 +9,5 @@ public interface ThreadPoolMetrics {
     void onTaskRejected();
     void onBpPaused(int delta);
     void onBpPauses();
-    int getRejectedTotal();
+    long getRejectedTotal();
 }
